src/GUI/requirements.txt
kivy
paho_mqtt==1.6.0
pillow